Replace FeedPresener with the generic presenter
