+++ ZZFeed/ZZFeed/Feed Presentation/FeedPresenter.swift
@@ -4,53 +4,12 @@
 
 import Foundation
 
-
-public struct FeedViewModel {
-    public let feed: [FeedItem]
-}
-
-public protocol FeedView {
-    func display(_ viewModel: FeedViewModel)
-}
-
 public final class FeedPresenter {
-    private let feedView: FeedView
-    private let loadingView: ResourceLoadingView
-    private let errorView: ResourceErrorView
-    
-    private var feedLoadError: String {
-        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
-                                 tableName: "Feed",
-                                 bundle: Bundle(for: FeedPresenter.self),
-                                 comment: "Error message displayed on connection problems.")
-    }
-    
     public static var title: String {
         return NSLocalizedString("FEED_VIEW_TITLE",
                                  tableName: "Feed",
                                  bundle: Bundle(for: FeedPresenter.self),
                                  comment: "Title for the feed view")
-    }
-    
-    public init(feedView: FeedView, loadingView: ResourceLoadingView, errorView: ResourceErrorView) {
-        self.feedView = feedView
-        self.loadingView = loadingView
-        self.errorView = errorView
-    }
-    
-    public func didStartLoadingFeed() {
-        errorView.display(.noError)
-        loadingView.display(ResourceLoadingViewModel(isLoading: true))
-    }
-    
-    public func didFinishLoadingFeed(with feed: [FeedItem]) {
-        loadingView.display(ResourceLoadingViewModel(isLoading: false))
-        feedView.display(FeedViewModel(feed: feed))
-    }
-    
-    public func didFinishLoadingFeed(with error: Error) {
-        errorView.display(.error(message: feedLoadError))
-        loadingView.display(ResourceLoadingViewModel(isLoading: false))
     }
     
     public static func map(_ models: [FeedItem]) -> FeedViewModel {
