+++ ZZFeed/ZZFeed/Feed Presentation/FeedView.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright © zzmasoud (github.com/zzmasoud).
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedItem]
+}
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
