+++ ZZFeed/ZZFeedTests/Feed Presentation/FeedPresentationTests.swift
@@ -11,57 +11,7 @@ class FeedPresentationTests: XCTestCase {
         XCTAssertEqual(FeedPresenter.title, localized("FEED_VIEW_TITLE"))
     }
     
-    func test_init_doesNotSendMessagesToView() {
-        let (_, view) = makeSUT()
-        
-        XCTAssertTrue(view.messages.isEmpty)
-    }
-    
-    func test_didStartLoadingFeed_displaysNoErrorMessageAndStartsLoading() {
-        let (sut, view) = makeSUT()
-
-        sut.didStartLoadingFeed()
-        
-        XCTAssertEqual(view.messages, [
-            .display(errorMessage: .none),
-            .display(isLoading: true)
-        ])
-    }
-    
-    func test_didFinishLoadingFeed_displaysFeedAndStopsLoading() {
-        let (sut, view) = makeSUT()
-        let feed = uniqueItems().models
-
-        sut.didFinishLoadingFeed(with: feed)
-        
-        XCTAssertEqual(view.messages, [
-            .display(feed: feed),
-            .display(isLoading: false)
-        ])
-    }
-    
-    func test_didFinishLoadingFeedWithError_displaysLocalizedErrorMessageAndStopsLoading() {
-        let (sut, view) = makeSUT()
-
-        sut.didFinishLoadingFeed(with: anyNSError())
-
-        XCTAssertEqual(view.messages, [
-            .display(errorMessage: localized("FEED_VIEW_CONNECTION_ERROR")),
-            .display(isLoading: false)
-        ])
-    }
-    
     // MARK: - Helpers
-    
-    private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: FeedPresenter, view: ViewSpy) {
-        let view = ViewSpy()
-        let sut = FeedPresenter(feedView: view, loadingView: view, errorView: view)
-        
-        trackForMemoryLeaks(view, file: file, line: line)
-        trackForMemoryLeaks(sut, file: file, line: line)
-        
-        return (sut, view)
-    }
     
     private func localized(_ key: String, file: StaticString = #file, line: UInt = #line) -> String {
         let table = "Feed"
@@ -71,27 +21,5 @@ class FeedPresentationTests: XCTestCase {
             XCTFail("Missing string in the table \(table) for key \(key)", file: file, line: line)
         }
         return value
-    }
-    
-    class ViewSpy: FeedView, ResourceLoadingView, ResourceErrorView {
-        enum Message: Hashable {
-            case display(errorMessage: String?)
-            case display(isLoading: Bool)
-            case display(feed: [FeedItem])
-        }
-        
-        private(set) var messages: Set<Message> = []
-        
-        func display(_ viewModel: ResourceErrorViewModel) {
-            messages.insert(.display(errorMessage: viewModel.message))
-        }
-        
-        func display(_ viewModel: ResourceLoadingViewModel) {
-            messages.insert(.display(isLoading: viewModel.isLoading))
-        }
-        
-        func display(_ viewModel: FeedViewModel) {
-            messages.insert(.display(feed: viewModel.feed))
-        }
     }
 }
